feat: 마감임박 아이템 등록, 조회 기능 추가

+++ src/services/queries/items/deserialize.ts
@@ -17,6 +17,6 @@ export const deserialize = (id: string, item: { [key: string]: string }): Item =
 		views: parseInt(item.views),
 		likes: parseInt(item.likes),
 		bids: parseInt(item.bids),
-		price: parseInt(item.price)
+		price: parseFloat(item.price)
 	};
 };
